abstracted out bounty combination
